+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to InAcademia SVS will be documented in this file.
 
 ## [Unreleased]
+## 2020-12-28
+### Added
+- Error handling for the case when SVS is unable to retrieve metadata against the resolved idp entity_id
+- Error handling for the case when SVS is unable to resolve SAML binding of the resolved idp entity_id
+
 ## 2020-12-08
 - Add idp_hint translation logic